feat: 화재 탐지 이슈 그룹핑 및 YOLO 재분석 제어 추가

+++ backend/src/main/java/com/burinake/dto/BoundingBox.java
@@ -1,10 +1,10 @@
 package com.burinake.dto;
 
 public record BoundingBox(
-        int x,
-        int y,
-        int width,
-        int height,
+        double x,
+        double y,
+        double width,
+        double height,
         String label,
         Double score
 ) {
